Finalizada toda a parte vizual e efeitos de CSS, faltando apenas implementar o efeito show more/show less
